requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aggdraw==1.3.19
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
argcomplete==3.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attr==0.3.1
attrs==23.2.0
audioread==3.0.1
azure-core==1.32.0
azure-storage-blob==12.25.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==5.0.1
blessings==1.7
blinker==1.9.0
boto==2.49.0
boto3==1.37.15
botocore==1.37.15
bqplot==0.12.44
branca==0.8.1
Brotli==1.1.0
cachetools==5.5.1
caller==0.0.2
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.2
constant==0.0.4
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
datamodel-code-generator==0.26.1
datasets==3.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
Django==5.1.7
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
earthengine-api==1.5.7
ee==0.2
eerepr==0.1.1
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
expiringdict==1.2.2
Faker==37.0.1
fastjsonschema==2.19.1
file_reader==0.2.3
filelock==3.16.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
folium==0.19.5
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
geemap==0.35.3
genson==1.3.0
geocoder==1.38.1
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.7.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.11.0
hilbertcurve==2.0.5
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.27.1
humanfriendly==10.0
humansignal-drf-yasg==1.21.10.post1
idna==3.7
ijson==3.3.0
imageio==2.37.0
imgviz==1.7.6
importlib_metadata==8.6.1
imutils==0.5.4
inflect==5.6.2
inflection==0.5.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.4
ipyleaflet==0.19.2
ipython==8.25.0
ipytree==0.2.2
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-leaflet==0.19.2
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
kagglehub==0.3.10
keras==3.5.0
kiwisolver==1.4.5
label-studio==1.16.0
label-studio-sdk==1.0.10
labelme==5.8.0
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lockfile==0.12.2
loguru==0.7.3
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mutagen==1.47.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.23.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.60.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.31.0
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.0.2
osam==0.2.3
osmnx==2.0.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
peewee==3.17.6
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==6.0.1
pooch==1.8.2
prometheus_client==0.20.0
prompt_toolkit==3.0.46
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.2
pygame==2.6.1
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.1.2
pyperclip==1.9.0
pyproj==3.7.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyRFC3339==2.0.1
pyshp==2.3.1
PySocks==1.7.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
python-slugify==8.0.4
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==26.0.3
ratelim==0.1.6
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.18.1
rq==1.16.2
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.11.4
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.5.1
scipy==1.14.0
scooby==0.10.0
seaborn==0.13.2
selenium==4.27.1
semver==3.0.4
Send2Trash==1.8.3
sentry-sdk==2.23.1
setuptools==76.1.0
shapely==2.0.7
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.0
soupsieve==2.5
soxr==0.5.0.post1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.41.1
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.18.0.post0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2025.3.13
tinycss2==1.3.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
traittypes==0.2.1
transformers==4.48.0
trio==0.28.0
trio-websocket==0.11.1
trl==0.13.0
tweepy==4.14.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
visualkeras==0.1.4
Wave==0.0.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.1.3
wheel==0.40.0
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.16.0
wsproto==1.2.0
xmljson==0.2.1
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
yfinance==0.2.43
yt-dlp==2024.8.6
zipp==3.21.0
